fix jsx problem for onboarding
