+++ src/index.js
@@ -10,4 +10,4 @@ export { useCollectionChunks } from "./useCollectionChunks";
 export { CustomChunkCollection } from "./CustomChunkCollection.jsx";
 export { ChunkCollectionContext } from "./ChunkCollectionContext.js";
 export { CollectionItemWrapper } from "./CollectionItemWrapper.jsx";
-export { renderOnboarding } from "./onboarding/index.js";
+export { renderOnboarding, Onboarding } from "./onboarding/index.jsx";
