--- src/onboarding/steps/coreConcepts/StepFive.js
@@ -1,93 +1,0 @@
-export default function StepThree() {
-    return (
-      <div className="relative py-16 bg-white overflow-hidden">
-        <div className="hidden lg:block lg:absolute lg:inset-y-0 lg:h-full lg:w-full">
-          <div className="relative flex justify-end h-full text-lg max-w-prose mx-auto" aria-hidden="true">
-            <svg
-              className="transform translate-x-32"
-              width={180}
-              height={180}
-              fill="none"
-              viewBox="0 0 404 384"
-            >
-              <defs>
-                <pattern
-                  id="74b3fd99-0a6f-4271-bef2-e80eeafdf357"
-                  x={0}
-                  y={0}
-                  width={20}
-                  height={20}
-                  patternUnits="userSpaceOnUse"
-                >
-                  <rect x={0} y={0} width={4} height={4} className="text-gray-200" fill="currentColor" />
-                </pattern>
-              </defs>
-              <rect width={404} height={384} fill="url(#74b3fd99-0a6f-4271-bef2-e80eeafdf357)" />
-            </svg>
-            <svg
-              className="absolute top-1/2 right-full transform -translate-y-1/2 -translate-x-32"
-              width={404}
-              height={384}
-              fill="none"
-              viewBox="0 0 404 384"
-            >
-              <defs>
-                <pattern
-                  id="f210dbf6-a58d-4871-961e-36d5016a0f49"
-                  x={0}
-                  y={0}
-                  width={20}
-                  height={20}
-                  patternUnits="userSpaceOnUse"
-                >
-                  <rect x={0} y={0} width={4} height={4} className="text-gray-200" fill="currentColor" />
-                </pattern>
-              </defs>
-              <rect width={404} height={384} fill="url(#f210dbf6-a58d-4871-961e-36d5016a0f49)" />
-            </svg>
-            <svg
-              className="absolute bottom-12 left-full transform translate-x-32"
-              width={404}
-              height={384}
-              fill="none"
-              viewBox="0 0 404 384"
-            >
-              <defs>
-                <pattern
-                  id="d3eb07ae-5182-43e6-857d-35c643af9034"
-                  x={0}
-                  y={0}
-                  width={20}
-                  height={20}
-                  patternUnits="userSpaceOnUse"
-                >
-                  <rect x={0} y={0} width={4} height={4} className="text-gray-200" fill="currentColor" />
-                </pattern>
-              </defs>
-              <rect width={404} height={384} fill="url(#d3eb07ae-5182-43e6-857d-35c643af9034)" />
-            </svg>
-          </div>
-        </div>
-        <div className="relative px-4 sm:px-6 lg:px-8">
-          <div className="text-lg max-w-prose mx-auto">
-            <h1>
-              <span className="block text-base text-center text-indigo-600 font-semibold tracking-wide uppercase">
-                Core Concept #2
-              </span>
-              <span className="mt-2 block text-3xl text-center leading-8 font-extrabold tracking-tight text-gray-900 sm:text-4xl">
-                Collections
-              </span>
-            </h1>
-          </div>
-          <div className="mt-6 prose prose-indigo prose-lg text-gray-500 mx-auto">
-            <p>
-            A collection is a group of chunks with the same fields. For example, you might have a "Team Member" collection, which has the fields First Name, Last Name, Title and Headshot.</p>
-            <br/>
-            <p>Fields can hold different types of information. For example, the First Name field above would hold text, whereas the Headshot field would hold an image.</p>
-            <br/>
-          </div>
-        </div>
-      </div>
-    )
-  }
-
