+++ src/onboarding/steps/editing/StepFive.jsx
