+++ src/onboarding/steps/editing/StepOne.jsx
